Fixes createSaving(); changed balance to double; started integrating stateful bean.

+++ ACME_BankingSystem/ACME_BankingSystem-ejb/src/java/Beans/Deposit_Saving.java
@@ -47,7 +47,7 @@ public class Deposit_Saving implements Deposit_SavingRemote {
     }
 
     @Override
-    public void InputBalance(Integer E_ID, String ACCNUM, Integer BALANCE, String desc) {
+    public void InputBalance(Integer E_ID, String ACCNUM, double BALANCE, String desc) {
         try {
             SavingDAO dao = new RDBSavingDAO(connection);
             Saving saving = new Saving(ACCNUM);
